+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to `ewwii` are documented here.
 This changelog follows the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format,
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [UNRELEASED]
+
+### Fixed
+
+- Dummy receiver not being stored for plugin buffer.
+
 ## [0.9.0] - 2026-06-13
 
 ### Added